chore(statblocks): quarantine legacy surfaces and repo hygiene (PR21)
Clarify v1 vs legacy authority, inventory consumers, and reduce repository
drag without changing the DungeonBuddy v1 contract.
