--- .VSCodeCounter/2024-12-17_21-13-16/diff-details.md
@@ -1,15 +1,0 @@
-# Diff Details
-
-Date : 2024-12-17 21:13:16
-
-Directory /media/drakosfire/Shared1/DungeonMind/DungeonMindServer
-
-Total : 0 files,  0 codes, 0 comments, 0 blanks, all 0 lines
-
-[Summary](results.md) / [Details](details.md) / [Diff Summary](diff.md) / Diff Details
-
-## Files
-| filename | language | code | comment | blank | total |
-| :--- | :--- | ---: | ---: | ---: | ---: |
-
-[Summary](results.md) / [Details](details.md) / [Diff Summary](diff.md) / Diff Details
