--- .VSCodeCounter/2024-12-17_21-13-16/diff.csv
@@ -1,2 +1,0 @@
-"filename", "language", "", "comment", "blank", "total"
-"Total", "-", , 0, 0, 0
