--- .VSCodeCounter/2024-12-17_21-16-02/diff.md
@@ -1,19 +1,0 @@
-# Diff Summary
-
-Date : 2024-12-17 21:16:02
-
-Directory /media/drakosfire/Shared1/DungeonMind/DungeonMindServer
-
-Total : 0 files,  0 codes, 0 comments, 0 blanks, all 0 lines
-
-[Summary](results.md) / [Details](details.md) / Diff Summary / [Diff Details](diff-details.md)
-
-## Languages
-| language | files | code | comment | blank | total |
-| :--- | ---: | ---: | ---: | ---: | ---: |
-
-## Directories
-| path | files | code | comment | blank | total |
-| :--- | ---: | ---: | ---: | ---: | ---: |
-
-[Summary](results.md) / [Details](details.md) / Diff Summary / [Diff Details](diff-details.md)
