--- .VSCodeCounter/2024-12-17_21-16-02/results.md
@@ -1,86 +1,0 @@
-# Summary
-
-Date : 2024-12-17 21:16:02
-
-Directory /media/drakosfire/Shared1/DungeonMind/DungeonMindServer
-
-Total : 147 files,  1920155 codes, 1158 comments, 4643 blanks, all 1925956 lines
-
-Summary / [Details](details.md) / [Diff Summary](diff.md) / [Diff Details](diff-details.md)
-
-## Languages
-| language | files | code | comment | blank | total |
-| :--- | ---: | ---: | ---: | ---: | ---: |
-| JSON | 9 | 1,893,083 | 0 | 1 | 1,893,084 |
-| CSS | 11 | 12,729 | 158 | 3,007 | 15,894 |
-| JavaScript | 45 | 4,190 | 377 | 583 | 5,150 |
-| Python | 44 | 3,447 | 309 | 576 | 4,332 |
-| Less | 13 | 2,878 | 264 | 132 | 3,274 |
-| toml | 1 | 2,849 | 0 | 220 | 3,069 |
-| HTML | 2 | 622 | 0 | 6 | 628 |
-| Markdown | 4 | 138 | 0 | 56 | 194 |
-| Shell Script | 4 | 89 | 34 | 40 | 163 |
-| XML | 6 | 55 | 0 | 1 | 56 |
-| JSON with Comments | 5 | 29 | 0 | 5 | 34 |
-| Ignore | 1 | 28 | 4 | 4 | 36 |
-| Docker | 1 | 14 | 12 | 11 | 37 |
-| Ini | 1 | 4 | 0 | 1 | 5 |
-
-## Directories
-| path | files | code | comment | blank | total |
-| :--- | ---: | ---: | ---: | ---: | ---: |
-| . | 147 | 1,920,155 | 1,158 | 4,643 | 1,925,956 |
-| . (Files) | 16 | 3,400 | 68 | 377 | 3,845 |
-| cardgenerator | 3 | 269 | 0 | 46 | 315 |
-| cloudflare | 2 | 42 | 4 | 12 | 58 |
-| cloudflareR2 | 3 | 54 | 38 | 20 | 112 |
-| firestore | 3 | 24 | 2 | 10 | 36 |
-| routers | 7 | 450 | 57 | 90 | 597 |
-| ruleslawyer | 6 | 1,892,807 | 18 | 31 | 1,892,856 |
-| static | 87 | 20,278 | 799 | 3,738 | 24,815 |
-| static/build | 6 | 45 | 4 | 0 | 49 |
-| static/build (Files) | 3 | 42 | 0 | 0 | 42 |
-| static/build/static | 3 | 3 | 4 | 0 | 7 |
-| static/build/static/css | 1 | 1 | 1 | 0 | 2 |
-| static/build/static/js | 2 | 2 | 3 | 0 | 5 |
-| static/css | 2 | 313 | 1 | 56 | 370 |
-| static/js | 2 | 2 | 3 | 0 | 5 |
-| static/root_store | 1 | 236 | 0 | 0 | 236 |
-| static/root_store/Enchanted_Roots_Gear_Emporium | 1 | 236 | 0 | 0 | 236 |
-| static/storegenerator | 76 | 19,682 | 791 | 3,682 | 24,155 |
-| static/storegenerator/css | 7 | 12,314 | 136 | 2,943 | 15,393 |
-| static/storegenerator/scripts | 15 | 1,631 | 351 | 321 | 2,303 |
-| static/storegenerator/themes | 54 | 5,737 | 304 | 418 | 6,459 |
-| static/storegenerator/themes (Files) | 2 | 66 | 4 | 1 | 71 |
-| static/storegenerator/themes/Legacy | 10 | 1,373 | 56 | 100 | 1,529 |
-| static/storegenerator/themes/Legacy/5ePHB | 10 | 1,373 | 56 | 100 | 1,529 |
-| static/storegenerator/themes/Legacy/5ePHB (Files) | 3 | 751 | 56 | 29 | 836 |
-| static/storegenerator/themes/Legacy/5ePHB/snippets | 7 | 622 | 0 | 71 | 693 |
-| static/storegenerator/themes/V3 | 26 | 3,395 | 196 | 263 | 3,854 |
-| static/storegenerator/themes/V3/5eDMG | 3 | 39 | 4 | 13 | 56 |
-| static/storegenerator/themes/V3/5ePHB | 14 | 1,965 | 98 | 178 | 2,241 |
-| static/storegenerator/themes/V3/5ePHB (Files) | 3 | 1,115 | 98 | 65 | 1,278 |
-| static/storegenerator/themes/V3/5ePHB/snippets | 11 | 850 | 0 | 113 | 963 |
-| static/storegenerator/themes/V3/Blank | 6 | 894 | 42 | 49 | 985 |
-| static/storegenerator/themes/V3/Blank (Files) | 3 | 836 | 42 | 37 | 915 |
-| static/storegenerator/themes/V3/Blank/snippets | 3 | 58 | 0 | 12 | 70 |
-| static/storegenerator/themes/V3/Journal | 3 | 497 | 52 | 23 | 572 |
-| static/storegenerator/themes/assets | 7 | 86 | 2 | 3 | 91 |
-| static/storegenerator/themes/codeMirror | 3 | 255 | 23 | 9 | 287 |
-| static/storegenerator/themes/codeMirror (Files) | 2 | 154 | 3 | 1 | 158 |
-| static/storegenerator/themes/codeMirror/customThemes | 1 | 101 | 20 | 8 | 129 |
-| static/storegenerator/themes/fonts | 6 | 562 | 23 | 42 | 627 |
-| static/storegenerator/themes/fonts/5e | 3 | 251 | 8 | 23 | 282 |
-| static/storegenerator/themes/fonts/5e legacy | 1 | 54 | 4 | 4 | 62 |
-| static/storegenerator/themes/fonts/Journal | 1 | 48 | 4 | 7 | 59 |
-| static/storegenerator/themes/fonts/icon fonts | 1 | 209 | 7 | 8 | 224 |
-| storegenerator | 4 | 667 | 37 | 67 | 771 |
-| tests | 16 | 2,164 | 135 | 252 | 2,551 |
-| tests (Files) | 3 | 82 | 12 | 26 | 120 |
-| tests/cloudflareR2 | 4 | 831 | 42 | 57 | 930 |
-| tests/firestore | 1 | 66 | 3 | 13 | 82 |
-| tests/rulesLawyer | 4 | 283 | 19 | 62 | 364 |
-| tests/session_managment | 1 | 140 | 20 | 27 | 187 |
-| tests/storegenerator | 3 | 762 | 39 | 67 | 868 |
-
-Summary / [Details](details.md) / [Diff Summary](diff.md) / [Diff Details](diff-details.md)
