+++ tests/statblockgenerator/test_statblockgenerator_v2_routes.py
@@ -5,7 +5,7 @@ from unittest.mock import AsyncMock
 from fastapi import FastAPI
 from fastapi.testclient import TestClient
 
-from routers import statblockgenerator_router
+from routers import statblock_v2_compatibility_router
 from routers.internal_auth import INTERNAL_KEY_ENV, INTERNAL_KEY_HEADER
 from statblockgenerator.models.statblock_models import StatBlockDetails
 
@@ -46,7 +46,7 @@ def sample_statblock():
 
 def client():
     app = FastAPI()
-    app.include_router(statblockgenerator_router.router)
+    app.include_router(statblock_v2_compatibility_router.router)
     return TestClient(app)
 
 
@@ -80,7 +80,7 @@ def test_generate_draft_accepts_basic_fixture(monkeypatch):
             },
         )
     )
-    monkeypatch.setattr(statblockgenerator_router.statblock_generator, "generate_creature", mock_generate)
+    monkeypatch.setattr(statblock_v2_compatibility_router.statblock_generator, "generate_creature", mock_generate)
 
     response = client().post("/api/statblockgenerator/v2/generate-draft", json=payload, headers=internal_headers())
 
@@ -99,7 +99,7 @@ def test_generate_draft_accepts_basic_fixture(monkeypatch):
 def test_render_draft_wraps_existing_statblock_without_generation(monkeypatch):
     monkeypatch.setenv(INTERNAL_KEY_ENV, "test-internal-key")
     mock_generate = AsyncMock()
-    monkeypatch.setattr(statblockgenerator_router.statblock_generator, "generate_creature", mock_generate)
+    monkeypatch.setattr(statblock_v2_compatibility_router.statblock_generator, "generate_creature", mock_generate)
     payload = {
         "request_id": "db-cmd-render-001",
         "statblock": sample_statblock().model_dump(by_alias=True),
@@ -142,7 +142,7 @@ def test_generate_draft_failure_returns_stable_error_envelope(monkeypatch):
     monkeypatch.setenv(INTERNAL_KEY_ENV, "test-internal-key")
     payload = json.loads((FIXTURE_DIR / "generate_from_prompt.basic.json").read_text())
     mock_generate = AsyncMock(return_value=(False, {"error": "OpenAI client not initialized"}))
-    monkeypatch.setattr(statblockgenerator_router.statblock_generator, "generate_creature", mock_generate)
+    monkeypatch.setattr(statblock_v2_compatibility_router.statblock_generator, "generate_creature", mock_generate)
 
     response = client().post("/api/statblockgenerator/v2/generate-draft", json=payload, headers=internal_headers())
